% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors.cool_warm_diverging}
\alias{colors.cool_warm_diverging}
\title{Cool–Warm Diverging Palette}
\usage{
colors.cool_warm_diverging()
}
\value{
A character vector of hexadecimal color codes representing the Cool–Warm Diverging palette.
}
\description{
This function returns a "Cool–Warm Diverging" palette for indicating deviation around a midpoint.
}
\examples{
cool_warm_diverging_colors <- colors.cool_warm_diverging()
plot(1:7, col = cool_warm_diverging_colors, pch = 19, cex = 3)

}
