% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors.earth_forest}
\alias{colors.earth_forest}
\title{Earth & Forest Palette}
\usage{
colors.earth_forest()
}
\value{
A character vector of hexadecimal color codes representing the Earth & Forest palette.
}
\description{
This function returns an "Earth & Forest" palette with natural, muted green and brown tones.
}
\examples{
earth_forest_colors <- colors.earth_forest()
plot(1:7, col = earth_forest_colors, pch = 19, cex = 3)

}
