% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors.grayscale}
\alias{colors.grayscale}
\title{Grayscale Color Palette}
\usage{
colors.grayscale()
}
\value{
A character vector of hexadecimal color codes representing the grayscale color palette.
}
\description{
This function returns a grayscale color palette ranging from white to black.
}
\examples{
grayscale_colors <- colors.grayscale()
plot(1:7, col = grayscale_colors, pch = 19, cex = 3)

}
