% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors.mixed}
\alias{colors.mixed}
\title{Mixed Color Palette}
\usage{
colors.mixed()
}
\value{
A character vector of hexadecimal color codes representing the mixed color palette.
}
\description{
This function returns a mixed color palette suitable for various visualizations.
}
\examples{
mixed_colors <- colors.mixed()
plot(1:7, col = mixed_colors, pch = 19, cex = 3)

}
