% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors.nighty}
\alias{colors.nighty}
\title{Nighty Color Palette}
\usage{
colors.nighty()
}
\value{
A character vector of hexadecimal color codes representing the nighty color palette.
}
\description{
This function returns a nighty color palette with deep, vibrant colors.
}
\examples{
nighty_colors <- colors.nighty()
plot(1:7, col = nighty_colors, pch = 19, cex = 3)

}
