% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors.ocean_depths}
\alias{colors.ocean_depths}
\title{Ocean Depths Palette}
\usage{
colors.ocean_depths()
}
\value{
A character vector of hexadecimal color codes representing the Ocean Depths palette.
}
\description{
This function returns an "Ocean Depths" palette with cool blue tones inspired by sea layers.
}
\examples{
ocean_depths_colors <- colors.ocean_depths()
plot(1:7, col = ocean_depths_colors, pch = 19, cex = 3)

}
