% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors.pastel_bloom}
\alias{colors.pastel_bloom}
\title{Pastel Bloom Palette}
\usage{
colors.pastel_bloom()
}
\value{
A character vector of hexadecimal color codes representing the Pastel Bloom palette.
}
\description{
This function returns a "Pastel Bloom" palette with soft, delicate spring colors.
}
\examples{
pastel_bloom_colors <- colors.pastel_bloom()
plot(1:7, col = pastel_bloom_colors, pch = 19, cex = 3)

}
