% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors.purples}
\alias{colors.purples}
\title{Purples Color Palette}
\usage{
colors.purples()
}
\value{
A character vector of hexadecimal color codes representing the purple color palette.
}
\description{
This function returns a purple color palette with various shades for use in visualizations.
}
\examples{
purple_colors <- colors.purples()
plot(1:7, col = purple_colors, pch = 19, cex = 3)

}
