% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors.red_to_green}
\alias{colors.red_to_green}
\title{Red to Green Color Palette}
\usage{
colors.red_to_green()
}
\value{
A character vector of hexadecimal color codes representing the red-to-green color palette.
}
\description{
This function returns a red-to-green color palette suitable for visualizing changes
or gradients between two extremes.
}
\examples{
red_to_green_colors <- colors.red_to_green()
plot(1:4, col = red_to_green_colors, pch = 19, cex = 3)

}
