% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors.sunset_glow}
\alias{colors.sunset_glow}
\title{Sunset Glow Palette}
\usage{
colors.sunset_glow()
}
\value{
A character vector of hexadecimal color codes representing the Sunset Glow palette.
}
\description{
This function returns a "Sunset Glow" palette with warm, smooth gradients from light peach to deep burgundy.
}
\examples{
sunset_glow_colors <- colors.sunset_glow()
plot(1:7, col = sunset_glow_colors, pch = 19, cex = 3)

}
