% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors.terra}
\alias{colors.terra}
\title{Terra Color Palette}
\usage{
colors.terra()
}
\value{
A character vector of hexadecimal color codes representing the terra color palette.
}
\description{
This function returns a terra color palette inspired by earthy tones.
}
\examples{
terra_colors <- colors.terra()
plot(1:7, col = terra_colors, pch = 19, cex = 3)

}
