% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors.vibrant_spectrum}
\alias{colors.vibrant_spectrum}
\title{Vibrant Spectrum Palette}
\usage{
colors.vibrant_spectrum()
}
\value{
A character vector of hexadecimal color codes representing the Vibrant Spectrum palette.
}
\description{
This function returns a "Vibrant Spectrum" palette with bright, high-contrast hues.
}
\examples{
vibrant_spectrum_colors <- colors.vibrant_spectrum()
plot(1:7, col = vibrant_spectrum_colors, pch = 19, cex = 3)

}
