% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoleaf.R
\name{.geoleaf.check_class}
\alias{.geoleaf.check_class}
\title{.geoleaf.check_class}
\usage{
.geoleaf.check_class(obj)
}
\arguments{
\item{obj}{An object to check.}
}
\value{
NULL if the check passes; otherwise, an error is raised.
}
\description{
Checks if an object is of class \code{geoleaf}.
}
