% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georef.R
\name{.georef.check_class}
\alias{.georef.check_class}
\title{.georef.check_class}
\usage{
.georef.check_class(obj)
}
\arguments{
\item{obj}{An object to check.}
}
\value{
NULL if the class is valid; otherwise, an error is raised.
}
\description{
Checks if the provided object is of class \code{georef}.
}
