% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{.pca.check_class}
\alias{.pca.check_class}
\title{Check if an object is of class 'pca'}
\usage{
.pca.check_class(this)
}
\arguments{
\item{this}{An object to be checked.}
}
\value{
NULL if the object is of class 'pca'; otherwise, an error is raised.
}
\description{
This function checks if the provided object is of class 'pca'.
}
\keyword{internal}
