% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{.pca.get_component_equations}
\alias{.pca.get_component_equations}
\title{Get equations for each principal component}
\usage{
.pca.get_component_equations(this)
}
\arguments{
\item{this}{An object of class 'pca'.}
}
\value{
A character vector containing equations for each principal component.
}
\description{
This function generates equations representing each principal component based on
their loadings.
}
