% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{.pca.invert_loads}
\alias{.pca.invert_loads}
\title{Invert loadings in PCA results}
\usage{
.pca.invert_loads(this)
}
\arguments{
\item{this}{An object of class 'pca'.}
}
\value{
The modified PCA object with inverted loadings if necessary.
}
\description{
This function ensures that the first principal component has positive loadings.
}
