% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{.pca.translate_component_names}
\alias{.pca.translate_component_names}
\title{Translate component names in PCA results}
\usage{
.pca.translate_component_names(obj)
}
\arguments{
\item{obj}{An object of class 'pca'.}
}
\value{
The modified PCA object with updated component names.
}
\description{
This function modifies the column names of principal components and loadings,
replacing 'PC' with 'CP'.
}
