% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaviz.R
\name{.pcaviz.check_class}
\alias{.pcaviz.check_class}
\title{Check if an Object is of Class 'pcaviz'}
\usage{
.pcaviz.check_class(this)
}
\arguments{
\item{this}{An object to be checked.}
}
\value{
NULL if the object is of class 'pcaviz'; otherwise, an error is raised.
}
\description{
This function checks if the provided object is of class 'pcaviz'.
}
\keyword{internal}
