% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaviz.R
\name{.pcaviz.from_ggplot}
\alias{.pcaviz.from_ggplot}
\title{Create a PCA Visualization Object from a ggplot Object}
\usage{
.pcaviz.from_ggplot(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{A ggplot object to be converted.}
}
\value{
An object of class 'pcaviz'.
}
\description{
This function converts a ggplot object into a PCA visualization object.
}
\examples{
pca_viz <- .pcaviz.from_ggplot(ggplot())

}
