% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{.ppt.check_class}
\alias{.ppt.check_class}
\title{Check PPT Class}
\usage{
.ppt.check_class(obj)
}
\arguments{
\item{obj}{The object to be checked.}
}
\value{
NULL if the object is valid; otherwise, an error is raised.
}
\description{
Verifies if the provided object is of class \code{ppt}.
}
