% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptpos.R
\name{.pptpos.check_class}
\alias{.pptpos.check_class}
\title{Check pptpos Class}
\usage{
.pptpos.check_class(obj)
}
\arguments{
\item{obj}{The object to check.}
}
\value{
NULL if the object is valid; otherwise, an error is raised.
}
\description{
Verifies if the provided object is of class \code{pptpos}.
}
