% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{.table.check_class}
\alias{.table.check_class}
\title{Check Table Class}
\usage{
.table.check_class(obj)
}
\arguments{
\item{obj}{An object to check.}
}
\value{
NULL if the check passes; otherwise, an error is raised.
}
\description{
Verifies if an object is of class \code{table}.
}
