% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizsize.R
\name{.vizsize.check_class}
\alias{.vizsize.check_class}
\title{Check Class of Visualization Size Object}
\usage{
.vizsize.check_class(obj)
}
\arguments{
\item{obj}{An object to be checked.}
}
\value{
NULL if the object is of class 'vizsize'; otherwise, an error is raised.
}
\description{
This function checks if the provided object is of class 'vizsize'.
}
\keyword{internal}
