% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizsize.R
\name{.vizsize.set_line_large}
\alias{.vizsize.set_line_large}
\title{Set Large Line Properties}
\usage{
.vizsize.set_line_large(obj)
}
\arguments{
\item{obj}{A visualization size object of class 'vizsize'.}
}
\value{
The modified visualization size object with large line properties.
}
\description{
This function sets large line properties for the visualization size object.
}
