% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizsize.R
\name{.vizsize.set_line_normal}
\alias{.vizsize.set_line_normal}
\title{Set Normal Line Properties}
\usage{
.vizsize.set_line_normal(obj)
}
\arguments{
\item{obj}{A visualization size object of class 'vizsize'.}
}
\value{
The modified visualization size object with normal line properties.
}
\description{
This function sets normal line properties for the visualization size object.
}
