% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizsize.R
\name{.vizsize.set_line_small}
\alias{.vizsize.set_line_small}
\title{Set Small Line Properties}
\usage{
.vizsize.set_line_small(this)
}
\arguments{
\item{this}{A visualization size object of class 'vizsize'.}
}
\value{
The modified visualization size object with small line properties.
}
\description{
This function sets small line properties for the visualization size object.
}
