% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizsize.R
\name{.vizsize.set_size_large}
\alias{.vizsize.set_size_large}
\title{Set Large Size Properties}
\usage{
.vizsize.set_size_large(obj)
}
\arguments{
\item{obj}{A visualization size object of class 'vizsize'.}
}
\value{
The modified visualization size object with large dimensions.
}
\description{
This function sets large dimensions for the visualization size object.
}
