% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizsize.R
\name{.vizsize.set_size_normal}
\alias{.vizsize.set_size_normal}
\title{Set Normal Size Properties}
\usage{
.vizsize.set_size_normal(this)
}
\arguments{
\item{this}{A visualization size object of class 'vizsize'.}
}
\value{
The modified visualization size object with normal dimensions.
}
\description{
This function sets normal dimensions for the visualization size object.
}
