% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizsize.R
\name{.vizsize.set_size_small}
\alias{.vizsize.set_size_small}
\title{Set Small Size Properties}
\usage{
.vizsize.set_size_small(this)
}
\arguments{
\item{this}{A visualization size object of class 'vizsize'.}
}
\value{
The modified visualization size object with small dimensions.
}
\description{
This function sets small dimensions for the visualization size object.
}
