% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizsize.R
\name{.vizsize.set_text_large}
\alias{.vizsize.set_text_large}
\title{Set Large Text Properties}
\usage{
.vizsize.set_text_large(this)
}
\arguments{
\item{this}{A visualization size object of class 'vizsize'.}
}
\description{
This function sets large text properties for the visualization size object.
}
