% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizsize.R
\name{.vizsize.set_text_normal}
\alias{.vizsize.set_text_normal}
\title{Set Normal Text Properties}
\usage{
.vizsize.set_text_normal(obj)
}
\arguments{
\item{obj}{A visualization size object of class 'vizsize'.}
}
\value{
The modified visualization size object with normal text properties.
}
\description{
This function sets normal text properties for the visualization size object.
}
