% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizsize.R
\name{.vizsize.set_text_small}
\alias{.vizsize.set_text_small}
\title{Set Small Text Properties}
\usage{
.vizsize.set_text_small(obj)
}
\arguments{
\item{obj}{A visualization size object of class 'vizsize'.}
}
\value{
The modified visualization size object with small text properties.
}
\description{
This function sets small text properties for the visualization size object.
}
