% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geogg.R
\name{geogg.add_labels}
\alias{geogg.add_labels}
\title{Add Labels to Geographical Plot}
\usage{
geogg.add_labels(this, labels, latitude, longitude)
}
\arguments{
\item{this}{An object of class 'geogg'.}

\item{labels}{A vector of labels for each point.}

\item{latitude}{A numeric vector for the latitudes of each point.}

\item{longitude}{A numeric vector for the longitudes of each point.}
}
\value{
The updated geogg object with labels added.
}
\description{
This function adds labels to specified points on a geographical ggplot.
}
\examples{
geogg_obj <- geogg() \%>\% geogg.add_labels(labels = c("A", "B"), latitude = c(0, 10), longitude = c(0, 10))
}
