% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geogg.R
\name{geogg.add_points}
\alias{geogg.add_points}
\title{Add Points to Geographical Plot}
\usage{
geogg.add_points(
  this,
  latitude,
  longitude,
  groups = NULL,
  groups_shape = NULL,
  color_map = NULL,
  shape_map = NULL,
  labels = NULL,
  labels_shape = NULL,
  legend_title = "Legend Title",
  add_new_scale = FALSE,
  point_size = 1
)
}
\arguments{
\item{this}{An object of class \code{geogg}.}

\item{latitude}{A numeric vector of latitudes for the points.}

\item{longitude}{A numeric vector of longitudes for the points.}

\item{groups}{An optional vector for grouping the points (fill), default is \code{NULL}.}

\item{groups_shape}{An optional vector for grouping the points shapes, default is \code{NULL}.}

\item{color_map}{A named character vector for color mapping of \code{groups}, default is \code{NULL}.}

\item{shape_map}{A named numeric vector for shape mapping of \code{groups_shape}, default is \code{NULL}.}

\item{labels}{An optional vector of labels for each fill group, default is \code{NULL}.}

\item{labels_shape}{An optional vector of labels for each shape group, default is \code{NULL}.}

\item{legend_title}{A character string for the legend title, default is \code{'Legend Title'}.}

\item{add_new_scale}{A logical value indicating if a new color scale should be added (useful after a surface layer), default is \code{FALSE}.}

\item{point_size}{A numeric value specifying the overall size of the points, default is \code{1}.}
}
\value{
The updated geogg object with points added.
}
\description{
This function adds points to a geographical ggplot object, with options for grouping and color mapping.
}
\examples{
geogg_obj <- geogg() \%>\% geogg.add_points(latitude = c(10, 20), longitude = c(10, 20), groups = c("A", "B"), color_map = c("A" = "red", "B" = "blue"))
}
