% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geogg.R
\name{geogg.add_surface}
\alias{geogg.add_surface}
\title{Add Surface Layer to Geographical Plot}
\usage{
geogg.add_surface(
  this,
  georef_obj,
  data,
  latitude,
  longitude,
  width = 100,
  height = 100,
  legend_title = "Surface Title",
  palette = colors.purples(),
  opacity = "BB",
  add_new_scale = FALSE
)
}
\arguments{
\item{this}{An object of class 'geogg'.}

\item{georef_obj}{A georef object containing georeferenced information.}

\item{data}{A numeric vector representing the data values for the surface.}

\item{latitude}{A numeric vector for the latitudes corresponding to the data.}

\item{longitude}{A numeric vector for the longitudes corresponding to the data.}

\item{width}{An integer specifying the raster width.}

\item{height}{An integer specifying the raster height.}

\item{legend_title}{A character string for the legend title, default is 'Surface Title'.}

\item{palette}{A character vector defining the color palette, default is colors.purples().}

\item{opacity}{A character specifying the opacity in hexadecimal (00-FF), default is 'BB'.}

\item{add_new_scale}{A logical value indicating if a new color scale should be added, default is FALSE.}
}
\value{
The updated geogg object with the surface layer added.
}
\description{
Adds a surface layer (such as a heatmap or raster) to the geographical ggplot for additional data visualization.
}
