% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geogg.R
\name{geogg.add_title}
\alias{geogg.add_title}
\title{Add Title to Geogg Plot}
\usage{
geogg.add_title(this, title)
}
\arguments{
\item{this}{A \code{geogg} object.}

\item{title}{A character string specifying the title to be added to the plot.}
}
\value{
The modified \code{geogg} object with the specified title added.
}
\description{
This function adds a title to a \code{geogg} plot.
}
