% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geogg.R
\name{geogg}
\alias{geogg}
\title{Initialize Geographical ggplot (geogg)}
\usage{
geogg(size = vizsize.parse("normal"))
}
\arguments{
\item{size}{A vizsize object or character to define the plot size, defaults to "normal".}
}
\value{
An object of class 'geogg' initialized with a clean theme for geographical data.
}
\description{
Creates a ggplot object for geographical visualizations with a specified size and theme.
}
