% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geogg.R
\name{geogg.theme_base}
\alias{geogg.theme_base}
\title{Apply Base Theme to Geographical ggplot}
\usage{
geogg.theme_base(this)
}
\arguments{
\item{this}{An object of class 'geogg'.}
}
\value{
The updated geogg object with base theme formatting applied.
}
\description{
Adds a base theme with formatted legend text to the geographical ggplot.
}
