% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geogg.R
\name{geogg.theme_clean}
\alias{geogg.theme_clean}
\title{Apply Clean Theme to Geographical ggplot}
\usage{
geogg.theme_clean(this)
}
\arguments{
\item{this}{An object of class 'geogg'.}
}
\value{
The updated geogg object with a clean theme applied.
}
\description{
Removes default ggplot elements to create a clean canvas for geographical data.
}
\examples{
geogg_obj <- geogg() \%>\% geogg.theme_clean()
}
