% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geogg.R
\name{geogg.without_legend}
\alias{geogg.without_legend}
\title{Remove Legend from geogg Plot}
\usage{
geogg.without_legend(this)
}
\arguments{
\item{this}{A \code{geogg} object.}
}
\value{
The updated \code{geogg} object with the legend removed.
}
\description{
This function removes the legend from a \code{geogg} plot, making it suitable for visualizations where the legend is not needed.
}
