% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoleaf.R
\name{geoleaf.add_legend_continuous}
\alias{geoleaf.add_legend_continuous}
\title{Add a Continuous Legend to a GeoLeaf Map}
\usage{
geoleaf.add_legend_continuous(
  this,
  data,
  title,
  position,
  palette = colors.purples()
)
}
\arguments{
\item{this}{A \code{geoleaf} object to which the legend will be added.}

\item{data}{A numeric vector used to determine the values for the legend.}

\item{title}{A character string specifying the title of the legend.}

\item{position}{A character string specifying the position of the legend. Must be one of
'bottomleft', 'bottomright', 'topleft', or 'topright'.}

\item{palette}{A character vector or function defining the color palette to be used in the legend.
Defaults to \code{colors.nighty()}.}
}
\value{
The updated \code{geoleaf} object with the added continuous legend.
}
\description{
Adds a continuous legend to a \code{geoleaf} map based on a numeric vector.
}
\details{
This function adds a continuous color legend to a \code{geoleaf} map using the provided numeric data.
The legend will display a color gradient based on the values in \code{data}. The \code{palette} argument
defines the color scheme, while the \code{position} argument controls where the legend is placed on the map.

The legend's breaks are calculated based on the minimum and maximum values of the \code{data} vector,
with five evenly spaced intervals. The function ensures that no missing data values (NA) are shown
by setting their color to transparent.
}
