% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoleaf.R
\name{geoleaf.add_legend_discrete}
\alias{geoleaf.add_legend_discrete}
\title{Add Discrete Legend to Map}
\usage{
geoleaf.add_legend_discrete(this, title, colors, labels, position)
}
\arguments{
\item{this}{A `geoleaf` map object.}

\item{title}{A character string specifying the title of the legend.}

\item{colors}{A character vector representing the colors for each discrete category in the legend.}

\item{labels}{A vector of labels corresponding to each color, describing the categories.}

\item{position}{A character string specifying the position of the legend on the map. Options are 'bottomleft', 'bottomright', 'topleft', or 'topright'.}
}
\value{
The modified `geoleaf` map object with the added legend.
}
\description{
This function adds a discrete legend to a `geoleaf` map, specifying a title, a set of colors, labels, and a position for the legend.
}
\details{
This function checks if the `title`, `colors`, and `position` parameters are character strings, and ensures that the length of `labels` matches the length of `colors`. If the lengths don't match, an error is raised. The legend is added to the map with the specified title, color set, labels, and position.
}
\examples{
geoleaf.map <- geoleaf() \%>\%
  geoleaf.add_legend_discrete(
    title = 'Category Legend',
    colors = c('#FF0000', '#00FF00', '#0000FF'),
    labels = c('Low', 'Medium', 'High'),
    position = 'bottomright'
  )

}
\seealso{
\code{\link{addLegend}}, \code{\link{geoleaf}}
}
