% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoleaf.R
\name{geoleaf.add_legend_pca}
\alias{geoleaf.add_legend_pca}
\title{Add PCA Performance Legend to Map}
\usage{
geoleaf.add_legend_pca(this)
}
\arguments{
\item{this}{A `geoleaf` map object.}
}
\value{
The modified `geoleaf` map object with the added legend.
}
\description{
This function adds a discrete legend to a `geoleaf` map representing the relative performance based on the first principal component (PC1) of a PCA analysis. The legend divides the values into quartiles.
}
\details{
The function adds a legend to the map with four discrete ranges representing relative performance:
\itemize{
  \item 'Menor CP1 |- 1º Quartil'
  \item '1º Quartil |- 2º Quartil'
  \item '2º Quartil |- 3º Quartil'
  \item '3º Quartil |-| Maior CP1'
}
The color gradient goes from red (lower values of PC1) to green (higher values of PC1), indicating the performance relative to the first principal component.
}
\examples{
geoleaf.map <- geoleaf() \%>\%
  geoleaf.add_legend_pca()

}
\seealso{
\code{\link{geoleaf.add_legend_discrete}}, \code{\link{colors.red_to_green}}
}
