% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoleaf.R
\name{geoleaf.add_legend_percentage_of_proficiency}
\alias{geoleaf.add_legend_percentage_of_proficiency}
\title{Add Percentage of Proficiency Legend to Map}
\usage{
geoleaf.add_legend_percentage_of_proficiency(
  this,
  subject = c("mathematics", "portuguese language")
)
}
\arguments{
\item{this}{A \code{geoleaf} map object.}

\item{subject}{A character string indicating the subject for which the legend should be displayed. Must be either 'mathematics' or 'portuguese language'. Defaults to 'mathematics'.}
}
\value{
The modified \code{geoleaf} map object with the added legend.
}
\description{
This function adds a discrete legend to a \code{geoleaf} map, representing the percentage of adequate learning in either Mathematics or Portuguese Language.
The legend ranges from 0\% to 100\% proficiency with corresponding color gradients.
}
\details{
This function checks if the \code{subject} is either 'mathematics' or 'portuguese language'.
It then adds a legend at the bottom-right corner of the map with four discrete ranges:
\itemize{
\item '0\% |- 25\%'
\item '25\% |- 50\%'
\item '50\% |- 70\%'
\item '70\% |-| 100\%'
}
The legend title will be either 'Aprendizado Adequado em Matemática' or 'Aprendizado Adequado em Língua Portuguesa', based on the selected subject.
}
