% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoleaf.R
\name{geoleaf.add_pca_points}
\alias{geoleaf.add_pca_points}
\title{geoleaf.add_pca_points}
\usage{
geoleaf.add_pca_points(
  this,
  pca_obj,
  latitude,
  longitude,
  labels = NULL,
  popups = NULL
)
}
\arguments{
\item{this}{A \code{geoleaf} object.}

\item{pca_obj}{A 'pca' object.}

\item{latitude}{A numeric vector of latitude coordinates for the PCA points.}

\item{longitude}{A numeric vector of longitude coordinates for the PCA points.}

\item{labels}{An optional vector of labels corresponding to each marker. If not provided, no labels are displayed.}

\item{popups}{An optional vector of popups corresponding to each marker. If not provided, no popups are displayed.}
}
\value{
The updated \code{geoleaf} object with added PCA points.

warning Points with missing location data will be omitted from the map.
}
\description{
Adds PCA points to the Leaflet map based on provided latitude and longitude.
}
\details{
This function filters out any points with missing location data before adding them to the map.
}
