% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoleaf.R
\name{geoleaf.add_points}
\alias{geoleaf.add_points}
\title{Add Points to Map}
\usage{
geoleaf.add_points(
  this,
  latitude,
  longitude,
  colors = NULL,
  labels = NULL,
  popups = NULL,
  point_size = 1
)
}
\arguments{
\item{this}{A `geoleaf` map object.}

\item{latitude}{A numeric vector specifying the latitude coordinates for the points.}

\item{longitude}{A numeric vector specifying the longitude coordinates for the points.}

\item{colors}{An optional character vector specifying the colors for each marker. If not provided, a default color is used.}

\item{labels}{An optional vector of labels corresponding to each marker. If not provided, no labels are displayed.}

\item{popups}{An optional vector of popups corresponding to each marker. If not provided, no popups are displayed.}

\item{point_size}{A numeric value indicating the size of the points on the map. Defaults to 1.}
}
\value{
The modified `geoleaf` map object with the added points.
}
\description{
This function adds circular markers at specified geographic coordinates (latitude and longitude) to a `geoleaf` map. Markers can be customized with colors, labels, optional popups, and an adjustable point size.
}
\details{
The function verifies that `latitude` and `longitude` are numeric vectors of the same length. If `labels` are not provided, they default to empty strings, and the marker radius is set to a smaller size. If `colors` are not provided, a default color is used for all markers. Popups, when provided, display additional information on each marker. The `point_size` parameter allows customization of the point size on the map. Points with missing coordinates are omitted with a warning indicating the number of points omitted.
}
\examples{
geoleaf.map <- geoleaf() \%>\%
  geoleaf.add_points(
    latitude = c(-23.5505, -22.9068),
    longitude = c(-46.6333, -43.1729),
    colors = c('#FF0000', '#00FF00'),
    labels = c('São Paulo', 'Rio de Janeiro'),
    popups = c('Capital of SP', 'Capital of RJ'),
    point_size = 2
  )

}
\seealso{
\code{\link{addCircleMarkers}}, \code{\link{geoleaf}}, \code{\link{leaflet}}
}
