% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoleaf.R
\name{geoleaf.add_surface}
\alias{geoleaf.add_surface}
\title{Add a surface layer to a geoleaf map}
\usage{
geoleaf.add_surface(
  this,
  georef_obj,
  data,
  latitude,
  longitude,
  palette = colors.purples(),
  width = 100,
  height = 100,
  add_legend = FALSE,
  legend_title = "Legend Title",
  legend_position = "bottomleft"
)
}
\arguments{
\item{this}{A \code{geoleaf} object to which the surface will be added.}

\item{georef_obj}{A \code{georef} object containing the geographic boundary that will be used for the raster.}

\item{data}{A numeric vector of values to be interpolated into the surface.}

\item{latitude}{A numeric vector of latitude coordinates corresponding to the \code{data} points.}

\item{longitude}{A numeric vector of longitude coordinates corresponding to the \code{data} points.}

\item{palette}{A character vector of colors to be used for the surface. Defaults to \code{colors.purples()}.}

\item{width}{The width of the raster in pixels. Defaults to 100.}

\item{height}{The height of the raster in pixels. Defaults to 100.}

\item{add_legend}{Logical, indicating whether to add a continuous legend for the surface. Defaults to \code{FALSE}.}

\item{legend_title}{A character string for the title of the legend. Defaults to 'Legend Title'.}

\item{legend_position}{A character string for the legend's position, e.g., 'bottomleft' or 'topright'. Defaults to 'bottomleft'.}
}
\value{
The modified \code{geoleaf} map object with the surface layer and an optional legend.
}
\description{
This function creates and adds a raster surface layer to a \code{geoleaf} map using spatial interpolation.
It's useful for visualizing continuous data, like socio-economic indicators, across a geographic area.
}
