% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoleaf.R
\name{geoleaf.add_tiles}
\alias{geoleaf.add_tiles}
\title{geoleaf.add_tiles}
\usage{
geoleaf.add_tiles(this)
}
\arguments{
\item{this}{A \code{geoleaf} object.}
}
\value{
The updated \code{geoleaf} object with added tiles.
}
\description{
Adds base tiles to the Leaflet map.
}
