% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoleaf.R
\name{geoleaf.pca_map}
\alias{geoleaf.pca_map}
\title{PCA Map}
\usage{
geoleaf.pca_map(
  pca_obj,
  latitude,
  longitude,
  labels = NULL,
  popups = NULL,
  add_boundary = FALSE,
  add_surface = FALSE,
  georef_obj = NULL,
  surface_data = NULL,
  surface_latitude = NULL,
  surface_longitude = NULL,
  surface_legend_title = "Legend Title",
  surface_palette = colors.purples(),
  surface_width = 100,
  surface_height = 100
)
}
\arguments{
\item{pca_obj}{A PCA object containing the PCA results.}

\item{latitude}{A numeric vector of latitude coordinates.}

\item{longitude}{A numeric vector of longitude coordinates.}

\item{labels}{An optional vector of labels for the points. Defaults to NULL.}

\item{popups}{An optional vector of popups corresponding to each marker. If not provided, no popups are displayed.}

\item{add_boundary}{Logical, indicating whether to add a geographic boundary from a georef object.}

\item{add_surface}{Logical, indicating whether to add a surface layer with interpolation.}

\item{georef_obj}{A georef object used for geographic boundaries or surface layers.}

\item{surface_data}{A numeric vector for the surface data.}

\item{surface_latitude}{A numeric vector for the latitude of the surface layer.}

\item{surface_longitude}{A numeric vector for the longitude of the surface layer.}

\item{surface_legend_title}{A character string for the surface layer's legend title.}

\item{surface_palette}{A character vector of colors to use for the surface layer.}

\item{surface_width}{Numeric value for the surface layer width.}

\item{surface_height}{Numeric value for the surface layer height.}
}
\value{
A \code{geoleaf} map object with added PCA points, boundaries, and optionally a surface layer.
}
\description{
This function creates a map that visualizes PCA (Principal Component Analysis) results using geographic coordinates.
}
