% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georef.R
\name{georef.from_geojson}
\alias{georef.from_geojson}
\title{georef.from_geojson}
\usage{
georef.from_geojson(geojson)
}
\arguments{
\item{geojson}{A character string representing the GeoJSON.}
}
\value{
A \code{georef} object containing the Simple Features (SF) representation of the GeoJSON.
}
\description{
Creates a \code{georef} object from a GeoJSON string.
}
