% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georef.R
\name{georef.from_points}
\alias{georef.from_points}
\title{georef.from_points}
\usage{
georef.from_points(latitude, longitude)
}
\arguments{
\item{latitude}{A numeric vector of latitude coordinates.}

\item{longitude}{A numeric vector of longitude coordinates.}
}
\value{
A \code{georef} object containing the Simple Features (SF) representation of the points.
}
\description{
Creates a \code{georef} object from latitude and longitude points.
}
