% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georef.R
\name{georef.from_sf}
\alias{georef.from_sf}
\title{georef.from_sf}
\usage{
georef.from_sf(sf_obj)
}
\arguments{
\item{sf_obj}{An SF object representing the geographical data.}
}
\value{
A \code{georef} object containing the transformed SF object with CRS set to 4326.
}
\description{
Creates a \code{georef} object from an SF object.
}
