% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georef.R
\name{georef.get_raster}
\alias{georef.get_raster}
\title{georef.get_raster}
\usage{
georef.get_raster(this, data, latitude, longitude, width, height)
}
\arguments{
\item{this}{A \code{georef} object.}

\item{data}{A numeric vector containing values to be interpolated.}

\item{latitude}{A numeric vector of latitude coordinates corresponding to the data points.}

\item{longitude}{A numeric vector of longitude coordinates corresponding to the data points.}

\item{width}{An integer specifying the width of the output raster.}

\item{height}{An integer specifying the height of the output raster.}
}
\value{
A terra SpatRaster object representing the interpolated values over the specified grid.
}
\description{
Generates a raster based on the provided data points and their geographical coordinates.
}
