% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georef.R
\name{georef.set_sf}
\alias{georef.set_sf}
\title{georef.set_sf}
\usage{
georef.set_sf(this, sf_obj)
}
\arguments{
\item{this}{A \code{georef} object.}

\item{sf_obj}{An SF object to be associated with the \code{georef}.}
}
\value{
The updated \code{georef} object with the SF data set.
}
\description{
Sets the Simple Features (SF) object for the \code{georef} instance.
}
