% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggviz.R
\name{ggviz.from_ggplot}
\alias{ggviz.from_ggplot}
\title{ggviz Constructor from a ggplot Object}
\usage{
ggviz.from_ggplot(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{A `ggplot` object to be converted.}
}
\value{
An object of class `ggviz`.
}
\description{
Creates an object of class `ggviz` from a `ggplot` chart.
}
