% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inep.R
\name{inep.abbreviate_school_names}
\alias{inep.abbreviate_school_names}
\title{Abbreviate School Names}
\usage{
inep.abbreviate_school_names(school_names, size = 2)
}
\arguments{
\item{school_names}{A character vector containing the names of the schools to be abbreviated.}

\item{size}{An integer specifying the maximum number of words to retain in the abbreviated name. Defaults to 2.}
}
\value{
A character vector of abbreviated school names with the first letter of each word capitalized.
}
\description{
This function abbreviates school names by retaining only the most meaningful words, based on a specified size limit. Commonly ignored words are filtered out to produce a concise representation of each school name.
}
\details{
The function converts all input school names to lowercase, splits them into individual words, and filters out predefined ignored words and numeric values. The remaining words are then concatenated based on the specified size limit.
}
\examples{
school_names <- c("Escola Municipal de Ensino Fundamental", "Colégio Dr. João da Silva")
abbreviated_names <- inep.abbreviate_school_names(school_names, size = 2)

}
