% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inep.R
\name{inep.get_percentage_of_proficiency_categories}
\alias{inep.get_percentage_of_proficiency_categories}
\title{Get Proficiency Categories}
\usage{
inep.get_percentage_of_proficiency_categories(values)
}
\arguments{
\item{values}{A numeric vector containing proficiency values to be categorized.}
}
\value{
A factor vector representing the proficiency categories: 'D', 'C', 'B', or 'A'.
}
\description{
This function categorizes numeric proficiency values into discrete categories based on predefined thresholds.
}
\details{
The function categorizes values as follows:
- 'D' for values less than 25
- 'C' for values between 25 and 49
- 'B' for values between 50 and 69
- 'A' for values 70 and above
}
\examples{
proficiency_values <- c(10, 30, 55, 80)
categories <- inep.get_percentage_of_proficiency_categories(proficiency_values)

}
