% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.filter}
\alias{pca.filter}
\title{pca.filter}
\usage{
pca.filter(this, mask)
}
\arguments{
\item{this}{A \code{pca} object containing PCA results.}

\item{mask}{A logical vector indicating which rows to keep.

return A filtered \code{pca} object with only the selected rows.

throws Error if the mask is not of type logical or does not match the number of rows in the data.}
}
\description{
Filters PCA data based on a logical mask.
}
\details{
This function filters the principal components and associated data in a
PCA object based on a logical vector (mask). Only rows corresponding to
TRUE values in the mask are retained.
}
