% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.from_data_frame}
\alias{pca.from_data_frame}
\title{Create a PCA object from a data frame}
\usage{
pca.from_data_frame(data, center = TRUE, scale = FALSE)
}
\arguments{
\item{data}{A data frame containing the data to be analyzed.}

\item{center}{A logical value indicating whether to center the variables.}

\item{scale}{A logical value indicating whether to scale the variables.}
}
\value{
An object of class 'pca' containing:
\item{standard_deviation}{Standard deviations of the principal components.}
\item{explained_variance}{Proportion of variance explained by each principal component.}
\item{loads}{Loadings of the variables on the principal components.}
\item{principal_components}{The principal component scores.}
\item{component_equations}{Equations representing each principal component.}
}
\description{
This function computes the principal components of a given data frame,
returning an object of class 'pca'.
}
\examples{
df <- data.frame(x = rnorm(100), y = rnorm(100), z = rnorm(100))
pca_result <- pca.from_data_frame(df)

}
