% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.get_categories}
\alias{pca.get_categories}
\title{pca.get_categories}
\usage{
pca.get_categories(this)
}
\arguments{
\item{this}{A \code{pca} object containing principal components.}
}
\value{
A factor indicating the category for each observation based on CP1.
}
\description{
Categorizes principal components into quartiles.
}
\details{
This function assigns categories (A, B, C, D) to the first principal component (CP1)
based on its quartile. Values are categorized as follows:
- D: below the 25th percentile
- C: between the 25th and 50th percentiles
- B: between the 50th and 75th percentiles
- A: above the 75th percentile
}
