% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.get_category_colors}
\alias{pca.get_category_colors}
\title{pca.get_category_colors}
\usage{
pca.get_category_colors(this, pallete = colors.red_to_green())
}
\arguments{
\item{this}{A \code{pca} object containing PCA results.}

\item{pallete}{A character vector representing a color palette.
Must have at least four colors.}
}
\value{
A character vector of colors corresponding to the PCA categories.

throws Error if the palette has less than four colors or is not of type character.
}
\description{
Retrieves colors for PCA categories based on a specified color palette.
}
\details{
This function assigns colors to PCA categories (A, B, C, D) using the provided
color palette. The palette must contain at least four colors.
}
