% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.get_ID}
\alias{pca.get_ID}
\title{Get the ID 'Indicador de Desigualdade' of a PCA object}
\usage{
pca.get_ID(this)
}
\arguments{
\item{this}{An object of class 'pca'.}
}
\value{
A numeric value representing the ID metric.
}
\description{
This function calculates an ID metric for the PCA object based on the centroid distance.
}
