% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.get_largest_variations}
\alias{pca.get_largest_variations}
\title{Get Largest Variations in PCA Data}
\usage{
pca.get_largest_variations(
  this,
  number,
  keys,
  years,
  labels = NULL,
  variation = "positive",
  errors = "warn"
)
}
\arguments{
\item{this}{A PCA object of class 'pca'.}

\item{number}{An integer specifying the number of largest variations to return.}

\item{keys}{A vector of keys (variables) for which the largest variations are to be calculated.}

\item{years}{A numeric vector indicating the years corresponding to the data points.}

\item{labels}{(Optional) A vector of labels for each key, used for annotation. Defaults to empty strings.}

\item{variation}{A character string indicating whether to calculate 'positive' or 'negative' variations.}

\item{errors}{A character string indicating how to handle errors:
'raise' to stop execution, 'warn' to issue a warning, or 'ignore' to suppress errors.}
}
\value{
A data frame containing the largest variations, including keys, labels,
and the first two principal component coordinates for the specified number of largest variations.
}
\description{
This function calculates the largest variations in principal component scores
between two specified years for given keys (variables). It helps in identifying
which variables have changed the most over time in PCA analysis, either positively
or negatively.
}
\details{
The function performs several checks on the input parameters to ensure they meet
the expected criteria. It calculates the difference in principal component scores
between the minimum and maximum years provided and returns a data frame with the
specified number of largest variations. If there are fewer variations than requested,
it will handle errors based on the specified \code{errors} parameter.
}
