% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca.sort}
\alias{pca.sort}
\title{Sort PCA Components and Data}
\usage{
pca.sort(this, by, descending = FALSE)
}
\arguments{
\item{this}{An object of the PCA class.}

\item{by}{A numeric vector of the same length as the number of rows in the PCA data. It specifies the ordering criteria.}

\item{descending}{A logical value indicating whether the sorting should be in descending order (default is `FALSE`).}
}
\value{
The modified PCA object with sorted principal components and data.
}
\description{
Sorts the principal components and the data matrix of a PCA object based on a specified vector.
}
\examples{
pca_object <- pca.sort(pca_object, by = some_vector, descending = TRUE)

}
