% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaviz.R
\name{pcaviz.add_ID}
\alias{pcaviz.add_ID}
\title{Add ID Annotation to PCA Visualization}
\usage{
pcaviz.add_ID(this)
}
\arguments{
\item{this}{An object of class 'pcaviz' that contains a PCA object.}
}
\value{
The modified 'pcaviz' object with the ID metric annotated on the plot.
}
\description{
This function adds an annotation displaying the ID metric of a PCA object
to a PCA visualization object.
}
\details{
The ID metric is calculated using the `pca.get_ID()` function.
The annotation is placed near the top right corner of the plot based on
the ranges of the first two principal components (CP1 and CP2).
}
\examples{
# Assuming 'pca_viz' is an existing pcaviz object with a PCA analysis
pca_viz <- pcaviz.add_ID(pca_viz)

}
