% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaviz.R
\name{pcaviz.add_largest_increases}
\alias{pcaviz.add_largest_increases}
\title{Add Largest Increases to PCA Scatter Plot}
\usage{
pcaviz.add_largest_increases(this, number, keys, years, labels = NULL)
}
\arguments{
\item{this}{A PCA visualization object of class 'pcaviz'.}

\item{number}{An integer specifying the number of largest increases to visualize.}

\item{keys}{A vector of keys (variables) for which the largest increases are to be identified.}

\item{years}{A vector of years corresponding to the data points being analyzed.}

\item{labels}{(Optional) A vector of labels for each key, used for annotation in the plot.
If NULL, defaults to ''.}
}
\value{
A modified PCA visualization object with segments and labels added for the largest increases.
}
\description{
This function identifies and visualizes the largest increases in a PCA scatter plot
by adding segments and labels for specified keys over a range of years.
It helps in understanding how certain variables have changed significantly
across the specified time frame.
}
\details{
The function first checks that the input object is valid. It then retrieves
the largest increases using the `pca.get_largest_increases` helper function.
The results are printed and subsequently visualized by adding segments and labels
to the PCA plot. The segments are colored using a red-to-green gradient to indicate
changes.
}
\examples{
# Assuming pca_viz is a valid PCA visualization object
pca_viz <- pcaviz.add_largest_increases(
  pca_viz,
  number = 5,
  keys = c("Variable1", "Variable2"),
  years = c(2020, 2021, 2022)
)

}
