% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaviz.R
\name{pcaviz.add_multi_group_points}
\alias{pcaviz.add_multi_group_points}
\title{Add Multi-Group Points to PCA Scatter Plot}
\usage{
pcaviz.add_multi_group_points(this, groups)
}
\arguments{
\item{this}{A PCA visualization object of class 'pcaviz'.}

\item{groups}{A factor indicating group membership for each point.}
}
\value{
The modified PCA visualization object with added group points.
}
\description{
This function adds points representing multiple groups to the PCA scatter plot.
}
