% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaviz.R
\name{pcaviz.add_title}
\alias{pcaviz.add_title}
\title{Add Title to PCA Visualization}
\usage{
pcaviz.add_title(this, title)
}
\arguments{
\item{this}{A \code{pcaviz} object representing the PCA visualization.}

\item{title}{A character string containing the title text to add to the visualization.}
}
\value{
The modified \code{pcaviz} object with the title added.
}
\description{
This function adds a title to a PCA visualization object, enhancing interpretability.
}
\details{
This function appends a specified title to an existing PCA plot, allowing for better context and understanding of the visualization. The title is typically displayed at the top of the plot.
}
