% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaviz.R
\name{pcaviz.from_ggplot}
\alias{pcaviz.from_ggplot}
\title{Create a PCA Visualization Object from a ggplot Object}
\usage{
pcaviz.from_ggplot(ggplot_obj, pca_obj = NULL)
}
\arguments{
\item{ggplot_obj}{A ggplot object that represents the PCA scatter plot.}

\item{pca_obj}{An object of class 'pca' containing the results of the PCA analysis.}
}
\value{
A 'pcaviz' object that combines the ggplot object with the PCA analysis.
}
\description{
This function converts a ggplot object into a PCA visualization object
by associating it with a PCA analysis object.
}
\details{
This function checks if the provided \code{pca_obj} is of class 'pca'
and ensures that \code{ggplot_obj} is indeed a ggplot object. The resulting
object can be used for further customization and visualization.
}
