% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaviz.R
\name{pcaviz.scatter}
\alias{pcaviz.scatter}
\title{Scatter Plot for PCA Visualization}
\usage{
pcaviz.scatter(
  pca_obj,
  labels = NULL,
  groups = NULL,
  include_ID = FALSE,
  size = vizsize()
)
}
\arguments{
\item{pca_obj}{A `pca` object containing the PCA results.}

\item{labels}{An optional vector of labels for the data points. Defaults to `NULL`.}

\item{groups}{An optional factor for grouping the data points. Defaults to `NULL`.}

\item{include_ID}{Logical, indicating whether to include the inequality indicator in the plot caption. Defaults to `FALSE`.}

\item{size}{A `vizsize` object or a numeric/text size parameter for customizing plot dimensions. Defaults to `vizsize()`.}
}
\value{
A `pcaviz` object representing the scatter plot.
}
\description{
Generates a scatter plot for visualizing the results of a Principal Component Analysis (PCA),
including options for grouping and labeling data points.
}
\details{
The scatter plot visualizes the first two principal components (CP1 and CP2) from the PCA analysis.
The function provides options to:
- Include data point labels.
- Group data points using distinct colors.
- Add the inequality indicator to the plot caption.

The axes are labeled with the percentage of variance explained by each principal component, and
the caption includes the equations for CP1 and CP2. Dashed horizontal and vertical lines are drawn at
the origin for reference.
}
\examples{
# Example usage:
pca_results <- perform_pca(data)
scatter_plot <- pcaviz.scatter(
  pca_obj = pca_results,
  labels = c("A", "B", "C"),
  groups = factor(c("Group1", "Group2", "Group1")),
  include_ID = TRUE
)

}
