% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaviz.R
\name{pcaviz.set_scale_scatter}
\alias{pcaviz.set_scale_scatter}
\title{Set Color Scale for PCA Scatter Plot}
\usage{
pcaviz.set_scale_scatter(this, groups)
}
\arguments{
\item{this}{An object of class 'pcaviz' that contains a PCA visualization.}

\item{groups}{A factor indicating the grouping of data points in the PCA scatter plot.}
}
\value{
The modified 'pcaviz' object with the specified color scale applied to the scatter plot.
}
\description{
This function sets a manual color scale for the scatter plot in a PCA visualization,
allowing different groups to be visually distinguished.
}
\details{
The function utilizes `scale_color_manual()` to assign colors to different
levels of the provided factor. The number of colors used corresponds to the number of
unique groups in the `groups` parameter.
}
\examples{
# Assuming 'pca_viz' is an existing pcaviz object and 'group_factor' is a factor
pca_viz <- pcaviz.set_scale_scatter(pca_viz, group_factor)

}
