% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaviz.R
\name{pcaviz.set_size}
\alias{pcaviz.set_size}
\title{Set Size Properties for PCA Visualization}
\usage{
pcaviz.set_size(this, size_obj = vizsize())
}
\arguments{
\item{this}{A PCA visualization object of class 'pcaviz'.}

\item{size_obj}{An object containing size properties (default is `vizsize()`).}
}
\value{
The modified PCA visualization object with updated size properties.
}
\description{
This function sets the size properties for a PCA visualization object.
}
