% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaviz.R
\name{pcaviz.set_theme_column}
\alias{pcaviz.set_theme_column}
\title{Set Theme for PCA Column Plot}
\usage{
pcaviz.set_theme_column(this)
}
\arguments{
\item{this}{A \code{pcaviz} object representing the PCA visualization.}
}
\value{
The modified \code{pcaviz} object with the custom theme applied.
}
\description{
Applies a standardized theme to a PCA column plot, enhancing visual consistency across elements like borders, grid lines, axis titles, and text sizes.
}
\details{
This function is designed to streamline the appearance of PCA column plots, setting specific sizes for axis titles, text, plot title, subtitle, and caption. It also adjusts margins for axis labels and removes unnecessary elements like minor grid lines and legends, providing a cleaner, more focused visualization.
}
