% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaviz.R
\name{pcaviz.set_theme_scatter}
\alias{pcaviz.set_theme_scatter}
\title{Set Theme for PCA Scatter Plot}
\usage{
pcaviz.set_theme_scatter(this)
}
\arguments{
\item{this}{A PCA visualization object of class 'pcaviz'.}
}
\value{
The modified PCA visualization object with updated theme.
}
\description{
This function sets a minimal theme for the PCA scatter plot.
}
