% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt.add_ggplot}
\alias{ppt.add_ggplot}
\title{Add ggplot to Slide}
\usage{
ppt.add_ggplot(this, ggplot_obj, position)
}
\arguments{
\item{this}{A \code{ppt} object.}

\item{ggplot_obj}{A ggplot object to add to the slide.}

\item{position}{A \code{pptpos} object specifying the position.}
}
\value{
The modified \code{ppt} object.
}
\description{
Adds a ggplot object to the current slide at a specified position.
}
