% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt.add_object}
\alias{ppt.add_object}
\title{Add Object to Slide}
\usage{
ppt.add_object(this, obj, position)
}
\arguments{
\item{this}{A \code{ppt} object.}

\item{obj}{The object to add to the slide.}

\item{position}{A \code{pptpos} object specifying the position.}
}
\value{
The modified \code{ppt} object.
}
\description{
Adds an object (e.g., table or ggplot) to the slide at a specified position.
}
