% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt.add_table}
\alias{ppt.add_table}
\title{Add Table to Slide}
\usage{
ppt.add_table(this, table_obj, position, fit_height = FALSE)
}
\arguments{
\item{this}{A \code{ppt} object.}

\item{table_obj}{A table object to add to the slide.}

\item{position}{A \code{pptpos} object specifying the position.}

\item{fit_height}{A logical indicating whether to fit the table height.}
}
\value{
The modified \code{ppt} object.
}
\description{
Adds a table to the slide at a specified position.
}
