% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt.add_text}
\alias{ppt.add_text}
\title{Add Text to Slide}
\usage{
ppt.add_text(this, text, position, font = "Calibri", size = 18, bold = FALSE)
}
\arguments{
\item{this}{A \code{ppt} object.}

\item{text}{A character string for the text.}

\item{position}{A \code{pptpos} object specifying the text position.}

\item{font}{A character string specifying the font family.}

\item{size}{A numeric value for the font size.}

\item{bold}{A logical value indicating whether the text is bold.}
}
\value{
The modified \code{ppt} object.
}
\description{
Adds text to the current slide at a specified position.
}
