% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt.add_title}
\alias{ppt.add_title}
\title{Add Title to Slide}
\usage{
ppt.add_title(this, title)
}
\arguments{
\item{this}{A `ppt` object.}

\item{title}{A character string for the slide title.}
}
\value{
The modified `ppt` object.
}
\description{
Adds a title to the current slide.
}
