% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt.add_transition}
\alias{ppt.add_transition}
\title{Add Transition Slide}
\usage{
ppt.add_transition(this, text)
}
\arguments{
\item{this}{A \code{ppt} object.}

\item{text}{A character string for the transition text.}
}
\value{
The modified \code{ppt} object.
}
\description{
Adds a transition slide with specified text in a predefined position.
}
