% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt.from_template}
\alias{ppt.from_template}
\title{Create a PPT Object from Template}
\usage{
ppt.from_template(template_path)
}
\arguments{
\item{template_path}{A character string specifying the path to the template PowerPoint file.}
}
\value{
A `ppt` object with the loaded template.
}
\description{
Initializes a `ppt` object from a specified PowerPoint template file.
}
