% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt.get_height}
\alias{ppt.get_height}
\title{Get Slide Height}
\usage{
ppt.get_height(this)
}
\arguments{
\item{this}{A `ppt` object.}
}
\value{
A numeric value representing the slide height.
}
\description{
Retrieves the height of the slides in the `ppt` object.
}
