% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt.get_horizontal_dimension}
\alias{ppt.get_horizontal_dimension}
\title{Get Horizontal Dimension}
\usage{
ppt.get_horizontal_dimension(this, relative_dimension)
}
\arguments{
\item{this}{A \code{ppt} object.}

\item{relative_dimension}{A numeric value (0 - 1) specifying the relative dimension.}
}
\value{
A numeric value for the absolute horizontal dimension.
}
\description{
Calculates the horizontal dimension relative to the slide width.
}
