% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt.get_layout}
\alias{ppt.get_layout}
\title{Get Slide Layout}
\usage{
ppt.get_layout(this)
}
\arguments{
\item{this}{A \code{ppt} object.}
}
\value{
A character string representing the layout name.
}
\description{
Retrieves the default layout name for blank slides in the template.
}
