% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt.get_length}
\alias{ppt.get_length}
\title{Get Number of Slides}
\usage{
ppt.get_length(this)
}
\arguments{
\item{this}{A `ppt` object.}
}
\value{
An integer representing the number of slides.
}
\description{
Retrieves the total number of slides in the `ppt` object.
}
