% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt.get_master}
\alias{ppt.get_master}
\title{Get Slide Master}
\usage{
ppt.get_master(this)
}
\arguments{
\item{this}{A `ppt` object.}
}
\value{
A character string representing the master name.
}
\description{
Retrieves the default master name for slides in the template.
}
