% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt.get_width}
\alias{ppt.get_width}
\title{Get Slide Width}
\usage{
ppt.get_width(this)
}
\arguments{
\item{this}{A \code{ppt} object.}
}
\value{
A numeric value representing the slide width.
}
\description{
Retrieves the width of the slides in the \code{ppt} object.
}
