% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt.move_slide}
\alias{ppt.move_slide}
\title{Move Slide}
\usage{
ppt.move_slide(this, index = NULL, to)
}
\arguments{
\item{this}{A \code{ppt} object.}

\item{index}{An optional integer specifying the current slide index.}

\item{to}{An integer specifying the target slide index.}
}
\value{
The modified \code{ppt} object.
}
\description{
Moves the current slide to a specified position within the slide order.
}
