% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt.new_slide}
\alias{ppt.new_slide}
\title{Add a New Slide}
\usage{
ppt.new_slide(this, title = NULL)
}
\arguments{
\item{this}{A \code{ppt} object.}

\item{title}{An optional character string specifying the title of the new slide.}
}
\value{
The modified \code{ppt} object.
}
\description{
Adds a new slide to the \code{ppt} object and sets it as the current slide.
}
