% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt.on_slide}
\alias{ppt.on_slide}
\title{Go to Slide}
\usage{
ppt.on_slide(this, index)
}
\arguments{
\item{this}{A \code{ppt} object.}

\item{index}{An integer specifying the slide index to set as current.}
}
\value{
The modified \code{ppt} object.
}
\description{
Sets the specified slide as the current slide in the \code{ppt} object.
}
