% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt.save}
\alias{ppt.save}
\title{Save PowerPoint}
\usage{
ppt.save(this, path)
}
\arguments{
\item{this}{A \code{ppt} object.}

\item{path}{A character string specifying the file path.}
}
\description{
Saves the \code{ppt} object as a PowerPoint file at the specified path.
}
