% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptpos.R
\name{pptpos.center}
\alias{pptpos.center}
\title{Center Position}
\usage{
pptpos.center()
}
\value{
A `pptpos` object with coordinates for a centered element.
}
\description{
Defines the position for a centered element in the slide with additional side margins.
}
