% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptpos.R
\name{pptpos.center_large}
\alias{pptpos.center_large}
\title{Center Large Position}
\usage{
pptpos.center_large()
}
\value{
A `pptpos` object with coordinates for a centered large element.
}
\description{
Defines the position for a large element in the center of the slide.
}
