% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptpos.R
\name{pptpos.document_title}
\alias{pptpos.document_title}
\title{Document Title Position}
\usage{
pptpos.document_title()
}
\value{
A \code{pptpos} object with coordinates for positioning a document title.
}
\description{
Defines the position for a document title in the slide, occupying the top section with increased padding.
}
