% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptpos.R
\name{pptpos.grid}
\alias{pptpos.grid}
\title{Grid Position}
\usage{
pptpos.grid(n_rows, n_columns, row, column, margin = 0.1)
}
\arguments{
\item{n_rows}{An integer specifying the total number of rows in the layout grid.}

\item{n_columns}{An integer specifying the total number of columns in the layout grid.}

\item{row}{An integer indicating the row in which the element should be positioned.}

\item{column}{An integer indicating the column in which the element should be positioned.}

\item{margin}{A numeric value for the margin around the element. Defaults to 0.1.}
}
\value{
A \code{pptpos} object with calculated positions for the specified grid cell.
}
\description{
Defines a custom grid-based position for an element within the slide layout.
}
