% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptpos.R
\name{pptpos.left_half}
\alias{pptpos.left_half}
\title{Left Half Position}
\usage{
pptpos.left_half()
}
\value{
A `pptpos` object with coordinates for positioning an element on the left half.
}
\description{
Defines the position for an element occupying the left half of the slide.
}
