% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptpos.R
\name{pptpos.right_half}
\alias{pptpos.right_half}
\title{Wide Left Position}
\usage{
pptpos.right_half()
}
\value{
A \code{pptpos} object with coordinates for a wide left-aligned element.
}
\description{
Defines the position for a wide element occupying the left side of the slide.
}
