% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptpos.R
\name{pptpos.title}
\alias{pptpos.title}
\title{Title Position}
\usage{
pptpos.title()
}
\value{
A `pptpos` object with coordinates for positioning a title.
}
\description{
Defines the position for a title element in a slide, taking up the top section.
}
