% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptpos.R
\name{pptpos.wide_right}
\alias{pptpos.wide_right}
\title{Wide Right Position}
\usage{
pptpos.wide_right()
}
\value{
A `pptpos` object with coordinates for a wide right-aligned element.
}
\description{
Defines the position for a wide element occupying the right side of the slide.
}
