% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats.inequality_indicator}
\alias{stats.inequality_indicator}
\title{Inequality Indicator}
\usage{
stats.inequality_indicator(data)
}
\arguments{
\item{data}{A numeric data frame or matrix with each row representing an observation.}
}
\value{
A numeric value representing the inequality indicator.
}
\description{
Calculates the inequality indicator for a given dataset, measuring the average distance
of each observation from the centroid in the data.
}
\examples{
data <- matrix(c(1, 2, 3, 4, 5, 6), nrow = 2)
inequality <- stats.inequality_indicator(data)

}
