% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table.fit_to_page}
\alias{table.fit_to_page}
\title{Fit Table to Page}
\usage{
table.fit_to_page(this, page_width, page_height = NULL)
}
\arguments{
\item{this}{A \code{table} object.}

\item{page_width}{A numeric value specifying the page width to fit the table within.}

\item{page_height}{An optional numeric value specifying the page height to fit the table within.}
}
\value{
The resized \code{table} object.
}
\description{
Adjusts the table size to fit within specified page dimensions.
}
