% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table.linear_model}
\alias{table.linear_model}
\title{Create a Flextable Summary of a Linear Regression Model}
\usage{
table.linear_model(model)
}
\arguments{
\item{model}{A fitted linear model object created using \code{lm}.}
}
\value{
A Flextable object containing a formatted summary of the regression model.
}
\description{
This function takes a fitted linear model object (\code{model}) as input and returns a formatted Flextable summarizing the regression results.
}
