% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table.padding}
\alias{table.padding}
\title{Set Padding for a Table}
\usage{
table.padding(this, padding, part = c("all", "body", "header", "footer"))
}
\arguments{
\item{this}{A `table` object.}

\item{padding}{The desired padding in points.}

\item{part}{The part of the table to modify the padding. Options include:
- `"all"`: Apply the padding to the entire table.
- `"body"`: Apply the padding to the table body.
- `"header"`: Apply the padding to the table header.
- `"footer"`: Apply the padding to the table footer.}
}
\value{
The modified `table` object.
}
\description{
Adjusts the padding of various parts of a `table` object.
}
\examples{
library(flextable)
data <- data.frame(A = 1:3, B = 4:6)
tbl <- table(data)
tbl <- table.padding(tbl, padding = 5, part = "all")

}
