% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table}
\alias{table}
\title{Table Constructor}
\usage{
table(dataframe, column_names = colnames(dataframe))
}
\arguments{
\item{dataframe}{A data frame containing the data for the table.}

\item{column_names}{A vector of column names for the table headers. Defaults to the column names of the \code{dataframe}.}
}
\value{
A \code{table} object styled with a dark theme.
}
\description{
Creates a \code{table} object, extending the \code{flextable} class for formatted tables.
}
