% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table.set_theme_dark}
\alias{table.set_theme_dark}
\title{Dark Theme for Table}
\usage{
table.set_theme_dark(this)
}
\arguments{
\item{this}{A \code{table} object.}
}
\value{
The styled \code{table} object with a dark theme applied.
}
\description{
Applies a dark theme to a \code{table} object.
}
