% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table.text_size}
\alias{table.text_size}
\title{Set Font Size for a Table}
\usage{
table.text_size(this, size, part = c("all", "body", "header", "footer"))
}
\arguments{
\item{this}{A \code{table} object.}

\item{size}{The desired font size in points.}

\item{part}{The part of the table to modify the font size. Options include:
\itemize{
\item \code{"all"}: Apply the font size to the entire table.
\item \code{"body"}: Apply the font size to the table body.
\item \code{"title"}: Apply the font size to the table title.
\item \code{"footer"}: Apply the font size to the table footer.
}}
}
\value{
The modified \code{table} object.
}
\description{
Adjusts the font size of various parts of a \code{table} object.
}
