% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{type.check_character}
\alias{type.check_character}
\title{type.check_character}
\usage{
type.check_character(obj, property = NULL)
}
\arguments{
\item{obj}{The object to check.}

\item{property}{Optional name of the property for error messages.}
}
\value{
NULL if the check passes; otherwise, an error is raised.
}
\description{
Checks if an object is of type character.
}
