% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{type.check_numeric}
\alias{type.check_numeric}
\title{type.check_numeric}
\usage{
type.check_numeric(obj, property = NULL)
}
\arguments{
\item{obj}{The object to check.}

\item{property}{Optional name of the property for error messages.}
}
\value{
NULL if the check passes; otherwise, an error is raised.
}
\description{
Checks if an object is of type numeric.
}
\details{
This function verifies that the provided object is classified as numeric.
}
