% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.get_percentage_of_proficiency_categories}
\alias{utils.get_percentage_of_proficiency_categories}
\title{Get Proficiency Categories}
\usage{
utils.get_percentage_of_proficiency_categories(values)
}
\arguments{
\item{values}{A numeric vector containing proficiency values to be categorized.}
}
\value{
A factor vector representing the proficiency categories: 'D', 'C', 'B', or 'A'.
}
\description{
This function categorizes numeric proficiency values into discrete categories based on predefined thresholds.
}
\details{
The function categorizes values as follows:
\itemize{
\item 'D' for values less than 25
\item 'C' for values between 25 and 49
\item 'B' for values between 50 and 69
\item 'A' for values 70 and above
}
}
\examples{
proficiency_values <- c(10, 30, 55, 80)
categories <- utils.get_percentage_of_proficiency_categories(proficiency_values)

}
