% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.get_percentage_of_proficiency_category_colors}
\alias{utils.get_percentage_of_proficiency_category_colors}
\title{Get Proficiency Category Colors}
\usage{
utils.get_percentage_of_proficiency_category_colors(
  values,
  palette = colors.red_to_green()
)
}
\arguments{
\item{values}{A numeric vector containing proficiency values for which colors will be assigned.}

\item{palette}{A character vector specifying the color palette to use. Defaults to a red-to-green gradient.}
}
\value{
A character vector of colors corresponding to the proficiency categories.
}
\description{
This function assigns colors to proficiency categories based on input numeric values and a specified color palette.
}
\details{
The function first categorizes the input values using \code{utils.get_percentage_of_proficiency_categories}. It then assigns a color to each category based on the provided palette, where:
\itemize{
\item 'D' corresponds to the first color in the palette
\item 'C' corresponds to the second color
\item 'B' corresponds to the third color
\item 'A' corresponds to the fourth color
}
}
\examples{
proficiency_values <- c(10, 30, 55, 80)
colors <- utils.get_percentage_of_proficiency_category_colors(proficiency_values)

}
