% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.magnitude}
\alias{utils.magnitude}
\title{Magnitude of Observations}
\usage{
utils.magnitude(data)
}
\arguments{
\item{data}{A numeric data frame or matrix with each row representing an observation.}
}
\value{
A numeric vector where each element corresponds to the magnitude of each observation.
}
\description{
Calculates the magnitude (Euclidean norm) of each observation in the dataset.
}
\examples{
data <- matrix(c(1, 2, 3, 4, 5, 6), nrow = 2)
magnitudes <- utils.magnitude(data)

}
