% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geogg.R
\name{.geogg.check_class}
\alias{.geogg.check_class}
\title{Check Geographical Plot Class}
\usage{
.geogg.check_class(obj)
}
\arguments{
\item{obj}{An object to be checked.}
}
\value{
NULL. This function is called for its side effect of error-checking.
}
\description{
This internal function verifies if the given object is of class `geogg`.
It raises an error if the object is not of the correct class.
}
\details{
The function ensures that the input `obj` is of type `geogg`. It is typically used within other `geogg` methods to validate input objects.
}
\examples{
\dontrun{
.geogg.check_class(my_geogg_obj)
}
}
\keyword{internal}
