% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geogg.R
\name{geogg.add_points}
\alias{geogg.add_points}
\title{Add Points to Geographical Plot}
\usage{
geogg.add_points(
  this,
  latitude,
  longitude,
  groups = NULL,
  color_map = NULL,
  labels = NULL,
  legend_title = "Legend Title",
  add_new_scale = FALSE,
  point_size = 1
)
}
\arguments{
\item{this}{An object of class 'geogg'.}

\item{latitude}{A numeric vector of latitudes for the points.}

\item{longitude}{A numeric vector of longitudes for the points.}

\item{groups}{An optional vector for grouping the points, default is NULL.}

\item{color_map}{A named character vector for color mapping groups, default is NULL.}

\item{labels}{An optional vector of labels for each group, default is NULL.}

\item{legend_title}{A character string for the legend title, default is 'Legend Title'.}

\item{add_new_scale}{A logical value indicating if a new color scale should be added, default is FALSE.}

\item{point_size}{The size of the points to be added into the plot.}
}
\value{
The updated geogg object with points added.
}
\description{
This function adds points to a geographical ggplot object, with options for grouping and color mapping.
}
\examples{
geogg_obj <- geogg() \%>\% geogg.add_points(latitude = c(10, 20), longitude = c(10, 20), groups = c("A", "B"), color_map = c("A" = "red", "B" = "blue"))
}
