% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geogg.R
\name{geogg.percentage_of_proficiency_map}
\alias{geogg.percentage_of_proficiency_map}
\title{Add Percentage of Proficiency Map to Geographical Plot}
\usage{
geogg.percentage_of_proficiency_map(
  data,
  subject = c("mathematics", "portuguese language"),
  latitude,
  longitude,
  labels = NULL,
  add_boundary = FALSE,
  add_surface = FALSE,
  georef_obj = NULL,
  surface_data = NULL,
  surface_latitude = NULL,
  surface_longitude = NULL,
  surface_legend_title = "Legend Title",
  surface_palette = colors.purples(),
  surface_width = 100,
  surface_height = 100,
  size = vizsize.parse("large")
)
}
\arguments{
\item{data}{A numeric vector representing the percentage of proficiency values.}

\item{subject}{A character string indicating the subject, either 'mathematics' or 'portuguese language'. Default is 'mathematics'.}

\item{latitude}{A numeric vector of latitudes for the proficiency points.}

\item{longitude}{A numeric vector of longitudes for the proficiency points.}

\item{labels}{An optional vector of labels for each point, default is NULL.}

\item{add_boundary}{A logical value to include a boundary overlay, default is FALSE.}

\item{add_surface}{A logical value to include a surface overlay, default is FALSE.}

\item{georef_obj}{A georef object for geographic reference, required if add_boundary or add_surface is TRUE.}

\item{surface_data}{A numeric vector representing surface data, required if add_surface is TRUE.}

\item{surface_latitude}{A numeric vector of latitudes for the surface data, required if add_surface is TRUE.}

\item{surface_longitude}{A numeric vector of longitudes for the surface data, required if add_surface is TRUE.}

\item{surface_legend_title}{A character string for the legend title of the surface overlay, default is 'Legend Title'.}

\item{surface_palette}{A character vector of colors for the surface overlay, default is a purple color palette.}

\item{surface_width}{A numeric value specifying the width of the surface grid, default is 100.}

\item{surface_height}{A numeric value specifying the height of the surface grid, default is 100.}

\item{size}{An object defining size specifications, default is 'large'.}
}
\value{
A geogg object with the proficiency map added.
}
\description{
This function creates a geographical plot that visualizes proficiency levels in a specified subject by adding points for each location, optionally with boundary and surface data.
}
\examples{
geogg_obj <- geogg.percentage_of_proficiency_map(data = c(20, 40, 60, 80), subject = 'mathematics', latitude = c(-15, -20), longitude = c(-47, -48))

}
