% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoleaf.R
\name{geoleaf.add_points}
\alias{geoleaf.add_points}
\title{Add Points to Map}
\usage{
geoleaf.add_points(
  this,
  latitude,
  longitude,
  colors = NULL,
  labels = NULL,
  popups = NULL,
  point_size = 1
)
}
\arguments{
\item{this}{A \code{geoleaf} map object.}

\item{latitude}{A numeric vector specifying the latitude coordinates for the points.}

\item{longitude}{A numeric vector specifying the longitude coordinates for the points.}

\item{colors}{An optional character vector specifying the colors for each marker. If not provided, a default color is used.}

\item{labels}{An optional vector of labels corresponding to each marker. If not provided, no labels are displayed.}

\item{popups}{An optional vector of popups corresponding to each marker. If not provided, no popups are displayed.}

\item{point_size}{A numeric value indicating the size of the points on the map. Defaults to 1.}
}
\value{
The modified \code{geoleaf} map object with the added points.
}
\description{
This function adds circular markers at specified geographic coordinates (latitude and longitude) to a \code{geoleaf} map. Markers can be customized with colors, labels, optional popups, and an adjustable point size.
}
\details{
The function verifies that \code{latitude} and \code{longitude} are numeric vectors of the same length. If \code{labels} are not provided, they default to empty strings, and the marker radius is set to a smaller size. If \code{colors} are not provided, a default color is used for all markers. Popups, when provided, display additional information on each marker. The \code{point_size} parameter allows customization of the point size on the map. Points with missing coordinates are omitted with a warning indicating the number of points omitted.
}
