% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoleaf.R
\name{geoleaf.add_surface}
\alias{geoleaf.add_surface}
\title{Add Continuous Legend to Map}
\usage{
geoleaf.add_surface(
  this,
  georef_obj,
  data,
  latitude,
  longitude,
  palette = colors.purples(),
  width = 100,
  height = 100,
  add_legend = FALSE,
  legend_title = "Legend Title",
  legend_position = "bottomleft"
)
}
\arguments{
\item{this}{A `geoleaf` map object.}

\item{data}{A numeric vector representing the data used to generate the color gradient for the legend.}

\item{palette}{A character string specifying the color palette to use for the gradient. Defaults to \code{colors.nighty()}.}

\item{title}{A character string specifying the title of the legend.}

\item{position}{A character string specifying the position of the legend on the map. Options are 'bottomleft', 'bottomright', 'topleft', or 'topright'.}
}
\value{
The modified `geoleaf` map object with the added continuous legend.
}
\description{
This function adds a continuous color gradient legend to a `geoleaf` map based on a numeric vector of data. The legend dynamically reflects the range of the data, with a title and customizable position.
}
\details{
The function creates a continuous color scale based on the input data, using the specified color palette. The legend is divided into five breaks, which represent the range of the data, and the color gradient is applied accordingly. The position and title of the legend are customizable.
}
\examples{
geoleaf.map <- geoleaf() \%>\%
  geoleaf.add_legend_continuous(
    data = c(1, 2, 3, 4, 5),
    title = 'Continuous Legend',
    position = 'bottomright'
  )

}
\seealso{
\code{\link{addLegend}}, \code{\link{colorNumeric}}, \code{\link{geoleaf}}
}
