% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggviz.R
\name{ggviz.from_ggplot}
\alias{ggviz.from_ggplot}
\title{Construtor de ggviz a partir de um objeto ggplot}
\usage{
ggviz.from_ggplot(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{Objeto `ggplot` a ser convertido.}
}
\value{
Um objeto da classe `ggviz`.
}
\description{
Cria um objeto da classe `ggviz` a partir de um gráfico `ggplot`.
}
