% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggviz.R
\name{ggviz.from_ggplot}
\alias{ggviz.from_ggplot}
\title{ggviz Constructor from a ggplot Object}
\usage{
ggviz.from_ggplot(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{A \code{ggplot} object to be converted.}
}
\value{
An object of class \code{ggviz}.
}
\description{
Creates an object of class \code{ggviz} from a \code{ggplot} chart.
}
