% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggviz.R
\name{ggviz.radar}
\alias{ggviz.radar}
\title{Construtor de gráfico de radar ggviz}
\usage{
ggviz.radar(
  data,
  colors = colors.mixed(),
  labels = NULL,
  title = NULL,
  show_score = FALSE,
  axes_to_invert = NULL,
  size = vizsize()
)
}
\arguments{
\item{data}{Um `data.frame` contendo os dados para o gráfico de radar.}

\item{colors}{Vetor de caracteres que define as cores dos grupos.}

\item{labels}{Vetor com os rótulos a serem usados para os grupos. Se `NULL`, será gerado automaticamente.}

\item{title}{Título do gráfico. Se `NULL`, nenhum título será exibido.}

\item{show_score}{Lógico. Se `TRUE`, exibe a pontuação relativa para o primeiro grupo.}

\item{axes_to_invert}{Vetor com os nomes dos eixos que devem ser invertidos (opcional).}

\item{size}{Objeto `vizsize` ou valores numéricos para definir os tamanhos do texto e linhas do gráfico.}
}
\value{
Um objeto da classe `ggviz` contendo o gráfico de radar.
}
\description{
Gera um gráfico de radar usando um data frame, com opções de personalização
de cores, títulos, rótulos e exibição de pontuação.
}
\examples{
\dontrun{
data <- data.frame(Group1 = c(70, 80, 90), Group2 = c(50, 60, 80))
ggviz.radar(data, colors = c("#FF0000", "#00FF00"))
}
}
