% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaviz.R
\name{pcaviz.add_labels}
\alias{pcaviz.add_labels}
\title{Add Labels to Points in PCA Scatter Plot}
\usage{
pcaviz.add_labels(this, labels)
}
\arguments{
\item{this}{A PCA visualization object of class 'pcaviz'.}

\item{labels}{A vector of labels corresponding to each point in the plot.}
}
\value{
The modified PCA visualization object with added labels.
}
\description{
This function adds text labels to points in the PCA scatter plot.
}
\examples{
pca_viz <- pcaviz.scatter(pca_obj, labels = c("Label1", "Label2", ...))
}
