% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaviz.R
\name{pcaviz.add_single_group_points}
\alias{pcaviz.add_single_group_points}
\title{Add Single Group Points to PCA Scatter Plot}
\usage{
pcaviz.add_single_group_points(this, color = colors.mixed()[1])
}
\arguments{
\item{this}{A PCA visualization object of class 'pcaviz'.}

\item{color}{Single color (name or hex) for all points when no groups are used.}
}
\value{
The modified PCA visualization object with added points.
}
\description{
This function adds points representing a single group to the PCA scatter plot.
}
