% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaviz.R
\name{pcaviz.scatter}
\alias{pcaviz.scatter}
\title{Create a Scatter Plot for PCA Results}
\usage{
pcaviz.scatter(
  pca_obj,
  labels = NULL,
  groups = NULL,
  include_ID = TRUE,
  size = vizsize()
)
}
\arguments{
\item{pca_obj}{A PCA object containing principal components.}

\item{labels}{Optional labels for the points (default is NULL).}

\item{groups}{Optional grouping factor for coloring points (default is NULL).}

\item{size}{Size properties for the plot (default is `vizsize()`).}
}
\value{
A ggplot object representing the PCA scatter plot.
}
\description{
This function generates a scatter plot of the principal components from a PCA object.
}
