% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt.R
\name{ppt.add_document_title}
\alias{ppt.add_document_title}
\title{Add Document Title}
\usage{
ppt.add_document_title(this, title, size = 40)
}
\arguments{
\item{this}{A \code{ppt} object.}

\item{title}{A character string for the document title.}

\item{size}{A numerical value for the size of the title.}
}
\value{
The modified \code{ppt} object.
}
\description{
Adds a large title to the current slide.
}
