% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptpos.R
\name{pptpos.parse}
\alias{pptpos.parse}
\title{Parse pptpos Object or Character Identifier}
\usage{
pptpos.parse(obj)
}
\arguments{
\item{obj}{A `pptpos` object or character identifier specifying a position preset.}
}
\value{
A `pptpos` object corresponding to the specified position.
}
\description{
Converts a `pptpos` object or character identifier into a `pptpos` position. Supported identifiers include
'center', 'center-large', 'left-half', 'right-half', 'wide-left', 'wide-right', 'title', and 'document-title'.
}
