% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table.add_header_row}
\alias{table.add_header_row}
\title{Add Header Row}
\usage{
table.add_header_row(this, column_names, column_widths)
}
\arguments{
\item{this}{A \code{table} object.}

\item{column_names}{A vector of column names for the header row.}

\item{column_widths}{A vector of column widths, corresponding to each column name.}
}
\value{
The modified \code{table} object with the new header row.
}
\description{
Adds a header row with specified column names and widths to a \code{table} object.
}
