% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table.padding}
\alias{table.padding}
\title{Set Padding for a Table}
\usage{
table.padding(this, padding, part = c("all", "body", "header", "footer"))
}
\arguments{
\item{this}{A \code{table} object.}

\item{padding}{The desired padding in points.}

\item{part}{The part of the table to modify the padding. Options include:
\itemize{
\item \code{"all"}: Apply the padding to the entire table.
\item \code{"body"}: Apply the padding to the table body.
\item \code{"header"}: Apply the padding to the table header.
\item \code{"footer"}: Apply the padding to the table footer.
}}
}
\value{
The modified \code{table} object.
}
\description{
Adjusts the padding of various parts of a \code{table} object.
}
