% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table}
\alias{table}
\title{Table Constructor}
\usage{
table(dataframe, column_names = colnames(dataframe))
}
\arguments{
\item{dataframe}{A data frame containing the data for the table.}

\item{column_names}{A vector of column names for the table headers. Defaults to the column names of the `dataframe`.}
}
\value{
A `table` object styled with a dark theme.
}
\description{
Creates a `table` object, extending the `flextable` class for formatted tables.
}
\examples{
data <- data.frame(A = 1:3, B = 4:6)
tbl <- table(data, column_names = c("Col A", "Col B"))

}
