% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table.set_theme_dark}
\alias{table.set_theme_dark}
\title{Dark Theme for Table}
\usage{
table.set_theme_dark(this)
}
\arguments{
\item{this}{A `table` object.}
}
\value{
The styled `table` object with a dark theme applied.
}
\description{
Applies a dark theme to a `table` object.
}
\examples{
data <- data.frame(A = 1:3, B = 4:6)
tbl <- table(data, column_names = c("Col A", "Col B"))
tbl <- table.set_theme_dark(tbl)

}
